LibraryManagement/requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Flask-Session==0.5.0
Flask-WTF==1.2.1
Werkzeug==3.0.1
wtforms==3.1.2
gunicorn==21.2.0